Added common utils file
