+++ cypress/support/utils/profile.js
@@ -1,18 +1,5 @@
-import { path } from "Texts/common";
 import { profileSelector } from "Selectors/profile";
 import { profileText } from "Texts/profile";
-
-export const navigateToProfile=()=>{
- cy.get(profileSelector.profileDropdown).invoke("show");
- cy.contains("Profile").click();
- cy.url().should("include", path.profilePath);
-};
-
-export const logout=()=>{
- cy.get(profileSelector.profileDropdown).invoke("show");
- cy.contains("Logout").click();
- cy.url().should("include",path.loginPath);
-};
 
 export const profilePageElements = () =>{
  for(const elements in profileSelector.profileElements ){
